% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{DEHU}
\alias{DEHU}
\title{Statistic of the DEH test based on a double estimation in PDE}
\usage{
DEHU(data, a)
}
\arguments{
\item{data}{a (d,n) numeric matrix containing the data.}

\item{a}{positive numeric number (tuning parameter).}
}
\value{
The value of the test statistic.
}
\description{
Computes the test statistic of the DEH based on a double estimation in PDE test.
}
\details{
This functions evaluates the teststatistic with the given data and the specified tuning parameter \code{a}.
Each row of the data Matrix contains one of the n (multivariate) sample with dimension d. To ensure that the computation works properly
\eqn{n \ge d+1} is needed. If that is not the case the test returns an error.
}
\references{
Dörr, P., Ebner, B., Henze, N. (2019) "A new test of multivariate normality by a double estimation in a characterizing PDE" \href{https://arxiv.org/abs/1911.10955}{arXiv:1911.10955}
}
