% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{empchar}
\alias{empchar}
\title{Berechnet das negative des gewichteten Abstandes zwischen der empirischen und theoretischen charakteristischen Funktion an der Stelle x}
\usage{
empchar(x, data, r)
}
\arguments{
\item{x}{d-dimensionaler Vektor}

\item{data}{n x d Matrix der skalierten Residuen}

\item{r}{reelle Konstante des PUTests}
}
\value{
reeller Wert
}
\description{
Berechnet das negative des gewichteten Abstandes zwischen der empirischen und theoretischen charakteristischen Funktion an der Stelle x
}
