% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{ENorm}
\alias{ENorm}
\title{Euclidian norm}
\usage{
ENorm(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A number.
}
\description{
A function that returns the euclidian norm of a vector.
}
\seealso{
Other helpfunctions: 
\code{\link{hilf.e}()},
\code{\link{standard}()}
}
\concept{helpfunctions}
