% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{H1}
\alias{H1}
\title{Hilfsfunktionen zur Berechnung der Teststatistik von Cox und Small}
\usage{
H1(data, vector)
}
\arguments{
\item{data}{n x d Matrix der skalierten Residuen}

\item{vector}{vector on the unit sphere}
}
\value{
reeller Wert
}
\description{
Hilfsfunktionen zur Berechnung der Teststatistik von Cox und Small
}
