% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{hilf.e}
\alias{hilf.e}
\title{Helpfunction for covariance kernel}
\usage{
hilf.e(x)
}
\arguments{
\item{x}{a numeric vector.}
}
\value{
A number.
}
\description{
A function that returns
}
\seealso{
Other helpfunctions: 
\code{\link{ENorm}()},
\code{\link{standard}()}
}
\concept{helpfunctions}
