% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{L2.random.field}
\alias{L2.random.field}
\title{L^2 random field}
\usage{
L2.random.field(d, n.iter = 1000, n.points = 100, a, kov)
}
\arguments{
\item{d}{a positive integer indicating the dimension.}

\item{n.iter}{a positive integer indicating the number of iterations.}

\item{n.points}{a positive integer specifing ???}

\item{a}{Positive numeric number.}

\item{kov}{a with d^(-2) normed covariance kernel.}
}
\value{
A number (the critical value).
}
\description{
Generates a L^2 random field
}
\details{
This function calculates the critical values given a covariance kernel.
}
