% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{MQ1}
\alias{MQ1}
\title{first statistic of Manzotti and Quiroz}
\usage{
MQ1(data)
}
\arguments{
\item{data}{a n x d matrix of d dimensional data vectors.}
}
\value{
Value of the test statistic
}
\description{
This function returns the value of the first statistic of Manzotti and Quiroz (2001).
}
\examples{
MQ1(MASS::mvrnorm(50,c(0,1),diag(1,2)))

}
\references{
Manzotti, A., and Quiroz, A.J. (2001), Spherical harmonics in quadratic forms for testing multivariate normality, Test, 10:87–104, \href{https://doi.org/10.1007/BF02595825}{DOI}
}
