% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnt-package.R
\name{print.mnt}
\alias{print.mnt}
\title{Print method for tests of multivariate normality}
\usage{
\method{print}{mnt}(x, ...)
}
\arguments{
\item{x}{object of class "mnt".}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
the argument x, invisibly, as for all \link{print} methods.
}
\description{
Printing objects of class "mnt".
}
\details{
A \code{mnt} object is a named list of numbers and character string, supplemented with \code{test} (the name of the teststatistic). \code{test} is displayed as a title.
The remaining elements are given in an aligned "name = value" format.
}
\examples{
print(test.DEHU(MASS::mvrnorm(50,c(0,1),diag(1,2)),a=1,MC=500))

}
