% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnt-package.R
\docType{data}
\name{Quantile095}
\alias{Quantile095}
\title{Simulated empirical 95\% quantiles of the tests contained in package \code{mnt}}
\format{
A data frame with 9 rows and 20 columns.
}
\usage{
Quantile095
}
\description{
A dataset containing the empirical 0.95 quantiles of the tests for the dimensions \code{d=2,3,5} and samplesizes \code{n=20,50,100} based on a Monte Carlo Simulation study with 100000 repetitions. The following parameters were used:
\itemize{
  \item For \code{\link{BHEP}} the parameter \code{a=1},
  \item for \code{\link{HV}} the parameter \code{a=5},
  \item for \code{\link{HJG}} the parameter \code{a=1.5},
  \item for \code{\link{HJM}} the parameter \code{a=1.5},
  \item for \code{\link{DEHT}} the parameter \code{a=0.25},
  \item for \code{\link{DEHU}} the parameter \code{a=0.5},
  \item for \code{\link{CS}} the parameter \code{Points=NULL},
  \item for \code{\link{PU}} the parameter \code{r=2},
  \item for \code{\link{MASkew}} the parameter \code{Points=NULL},
  \item for \code{\link{MAKurt}} the parameter \code{Points=NULL},
}
}
\keyword{datasets}
