% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{standard}
\alias{standard}
\title{Standarising data}
\usage{
standard(data)
}
\arguments{
\item{data}{a n x d matrix. ???}
}
\value{
A n x d matrix.
}
\description{
A function that return the standarized data.
}
\seealso{
Other helpfunctions: 
\code{\link{ENorm}()},
\code{\link{hilf.e}()}
}
\concept{helpfunctions}
