% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpFunctions.R
\name{standard}
\alias{standard}
\title{Empirical scaled residuals}
\usage{
standard(data)
}
\arguments{
\item{data}{a n x d matrix of d dimensional data vectors..}
}
\value{
A n x d matrix of the scaled residuals.
}
\description{
A function that computes the scaled residuals of the data.
}
